pyproject.toml
[tool.poetry]
name = "probably"
version = "3.0.0"
description = ""
authors = ["Philipp Schröer <mail@philworld.de>"]

[tool.poetry.dependencies]
python = ">=3.7"
lark-parser = "^0.8.5"
algebraic-data-types = "^0.1.3"
attrs = "^19.3.0"
pysmt = "^0.9.0"
click = "^7.1.2"
sphinxcontrib-katex = "^0.7.1"
graphviz = "^0.15"
networkx = "2.5"
sympy = "^1.8"
matplotlib = "^3.4.2"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
pylint = "^2.5.2"
mypy = "^0.770"
yapf = "^0.30.0"
sphinx-autodoc-typehints = "^1.10.3"
sphinx-bootstrap-theme = "^0.7.1"
sphinx-click = "^2.3.2"
sphinx-git = "^11.0.0"
rope = "^0.17.0"
pytest-cov = "^2.9.0"
hypothesis = "^5.18.0"


[tool.poetry.scripts]
probably = "probably.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
